refactor: NameValidator로 메서드 이동
- NameValidator는 싱글톤 패턴을 이용해 객체 하나만을 사용하여 자원을 절약함.
- 해당 기능에서 사용되는 상수는 NameConstant라는 클래스로 분리해서 관리.
