+++ src/main/java/racingcar/NameConstant.java
@@ -1,0 +1,6 @@
+package racingcar;
+
+public class NameConstant {
+    public static final int NAME_MAX_LENGTH=5;
+    public static final int NAME_EMPTY_SIZE=0;
+}
