+++ src/test/java/racingcar/RacingGameTest.java
@@ -30,14 +30,4 @@ public class RacingGameTest {
         assertThatThrownBy(()->racingGame.checkAttemptCountIsInteger(input))
                 .isInstanceOf(IllegalArgumentException.class);
     }
-
-    @DisplayName("이름의 적합성을 판단하는 기능")
-    @ParameterizedTest
-    @ValueSource(strings={"eleven","abcdef","qwerdf","111234","1q2w3e4r","","  "})
-    void validateName_이름의_적합성_판단(String name){
-        RacingGame racingGame=new RacingGame();
-
-        assertThatThrownBy(()->racingGame.validateName(name))
-                .isInstanceOf(IllegalArgumentException.class);
-    }
 }
